sysdir: fix formatting error message with NULL string
When trying to the template dir, we pass in a `NULL` pointer for the
filename. That's perfectly fine, but if we're failing to find the
template directory then we'll creat an error message with the `NULL`
pointer passed in.

Fix the issue by setting different error messages based on whether the
filename is given or not. This even makes sense, as we're not searching
for a file in case we have no `name`, but for a directory. So the error
would've been misleading anyway.