errors: add GIT_EMISSING
"Missing" means that an object isn't present in the ODB but should
be available from a promisor remote or similar. Promisor packfiles may
reference objects that aren't present in the packfile but are available
remotely.

Compare with to GIT_ENOTFOUND which indicates there is no information to
suggest the object will ever be available - the repository may be
corrupt or a oid lookup is not present.