odb_pack: return EMISSING instead of ENOTFOUND for promisor packfiles.
If an object is referenced from a promisor packfile and the object is
not found elsewhere in the ODB, then return EMISSING instead of
ENOTFOUND.

Promisor packfiles can refer to objects that are not available in the
pack but should be available remotely. Contrast with ENOTFOUND to
indicate there is no information this object will ever exist.