Split the CI tests' steps
Previously, if a CI run failed, one would need to go hunt for the step
that failed among the very long error logs.

This test now runs each test step in its own little `step`, so that the
error for a particular test is always displayed at the very end and is
easy to find.