Implements .promisor packfile detection
A .pack file can be marked as being "keep" or "promisor" if a file
exists next to it with the same name but a different extension.
Before this change, .keep files were being checked but .promisor were
not. Now both are being checked (but neither one is used for anything).