Only avoid `mmap(2)`/`ftruncate(2)` when in non-Windows
It turns out that if we use `mmap(2)`, non-Windows remote filesystems
break due to permissions. If we don't, _Windows_ remote filesystems
break due to lack of coherence between memory mapped views of the file
and direct I/O operations done to the files.

To break out of this impossible situation, conditionally-compile
versions of Windows-specific `write_at` and `append_to_pack`.