attr_file: don't validate workdir paths in attr lookups
When looking up attributes for a file, we construct an absolute path
to the queried file within the working directory so that we can accept
both absolute paths and working directory relative paths.  We then trim
the leading working directory path to give us an in-repo path.

Since we only want the in-repo path to look up attributes - and not to
read it from disk - we don't need to validate its length.