xdiff: use xdl_free not free
We've added a lovely abstraction layer in xdiff so that it can call our
allocation functions.  But it also needs to call our free functions.  We
missed some `free` calls in `xmerge.c`.  Update them to use `xdl_free`.
Without this, we will pass a pointer allocated with a custom allocator
to the system free function.  :bomb: