util: add fallback for non-standard getloadavg()
The getloadavg() function is not part of the POSIX specification. While it is
implemented on *BSD and Linux, it is not available on the Haiku platform.
This causes linking errors when using libgit2. This commit adds a check if
the function is available while building for the target platform, and if it
is not, it will fall back to drand48() which is part of the POSIX standard
and should be available on all platforms.

While the drand48() function does not provide the same guarantee of
randomness as using system data, it does allow libgit2 continue to work on
more platforms.

This change also reintroduces the check for getentropy(), which seems to have
fallen through the cracks in the refactoring to the structure done in ef4ab29
and c3b7ace.

Finally, when building on Haiku, it now links to libgnu.so, which provides
the qsort_r symbol.