Do not register fake repository as owner of ODB
Accessing ODB is guaranteed to be thread-safe. Registering a
fake repository (created via git_repository_wrap_odb()) may
not register itself as owner of the ODB in order to maintain
that guarantee. Otherwise, accessing objects from ODB will
try to obtain the cache from the owning repository (via
odb_cache()) and produce a race if this ODB is concurrently
used to create fake repositories in other threads.
Consequently, operations on fake repositories will interact
with the ODB's cache instead of the repository's cache.