chunk-format: create an API from common code
The commit-graph and multi-pack-index file formats use common
substructures. This is so much the case that commit_graph.c and midx.c
have some common code that can be extracted into its own API.

Start by pulling out the write_offset() and write_chunk_header()
methods. Later changes will update the API with the reading half of the
API.

There are similar changes in the Git project in
git/git:chunk-format.[c|h] that could be used as an example for future
abstractions, such as populating a struct with the full list of chunks,
then letting the chunk API write each of the chunks using callbacks to
the specifics within each chunk.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>