repo: don't stat() safe.directory config entries
We need to turn `safe.directory` config entries that were specified into
our internal representation of directories, which always have a trailing
slash. This is to compare with our repository path(s) that already have
a trailing slash.

Use `path_to_dir` which appends a slash (if it's missing) instead of
`prettify` which canonicalizes the directory. This is both an expensive
operation (particularly on network drives or missing drives, where a
call may end up timing out) and incorrect, as git compares the literal
data in the config file against the canonicalized repository path.