push: set generic error in push_negotiation cb
When a user returns `-1` in a `push_negotiation` callback, we set the
error to whatever's hanging out in the buffer, probably something about
a missing configuration entry.

Clear the error buffer before invoking the callback, so that if a user
does not set an error message in their callback that we can detect. If
there is no error but `-1` is returned, set a generic error message.