clone: don't swallow error in should_checkout
When determining whether to check out a branch, we did not actually look
at `should_checkout`'s possible error condition. Disconnect the boolean
"should checkout" value from the potential error code so that it no
longer looks like a function that can be treated as if it returns a
boolean.