config: return only backends that contain the key
When deleting a key from a repository with multiple config
backends (like a --local config and a --worktree config), it's
important to return the correct backend to modify.

This patch ensures that we don't return a backend that is incapable of
deleting a given piece of configuration when that is the required use.