WIP proposal
In this version:

1. If there is exactly one config backend, use it.
2. If `--local` is available, use it.
3. Signal an error (as `git` does).

This closely simulates the behavior of git-config.

A test does fail, however, introduced in [1]:

    test_config_readonly__writing_to_cfg_with_ro_precedence_succeeds

Reading the message, I think the spirit of the readonly functionality
is still satisfied and the test can be adjusted.

Another test also fails:

    test_remote_httpproxy__env

This one eludes me -- because it only fails *sometimes*. It does not
fail if I just run that suite or surrounding suites. When it does
fail, it fails on the very first assertion:

    /* HTTP proxy is ignored for HTTPS */
    cl_setenv("HTTP_PROXY", "http://localhost:9/");
    assert_proxy_is(NULL);

I'm not sure what kind of test interaction could be going on here, but
I suspect (given my changes) that the problem is somewhere within
`http_proxy_config`.

[1]: 95f29fb35b9fbc7bd98f20bd91886522ddc03f4b