repo: support the preciousObjects extension
libgit2 implicitly supports precious objects, since there's no gc
command, nor even any option in our object database functionality to
delete an object from the odb.

In the future, when we support deleting objects, or a gc functionality,
we will need to honor the preciousObjects extension and reject those
APIs when it is set. In the meantime, since users cannot use libgit2
(directly) to delete an object, we can simply add this to our allowlist.