diff: always print the file header on GIT_DIFF_FORMAT_PATCH_HEADER
In `diff_print_patch_file` we try to avoid printing the file if we're not sure
that there will be some content. This works fine if we have any later functions
that might get called as part of printing. But when given the format
`GIT_DIFF_FORMAT_PATCH_HEADER`, this is the only function to get called.

Force us to print the diff header for the file in such a situation. This may
still print the header where e.g. `GIT_DIFF_FORMAT_PATCH` wouldn't, but it's
better than this option not working at all.